nutrimenu-api/requirements.txt
dotenv~=0.9.9
python-dateutil~=2.9.0

fastapi[all]~=0.116.1

google-genai
google-cloud-firestore

firebase-admin~=7.0.0